Fix mouse click to position cursor - handle clicks at view level
The initial implementation tried to handle mouse clicks within the EditorElement itself,
but this approach had critical issues:
1. The EditorElement didn't have access to the view context needed for cx.notify()
2. The canvas element was likely blocking mouse events from propagating
3. Without notify(), the view wasn't rerendering after cursor position changes

Solution:
- Moved mouse click handling to the EditorView level in the demo
- Uses cx.listener() to properly access the view's context
- Can now call cx.notify() to trigger a re-render after cursor moves
- Also focuses the editor when clicked

The click handler:
- Calculates the clicked position relative to the editor bounds
- Accounts for gutter width and scroll offset
- Converts pixel coordinates to row/column positions
- Clamps positions to valid buffer bounds
- Updates cursor position and ensures it's visible

This properly implements click-to-position functionality that actually works.
